feat: common notification component
