+++ frontend/package.json
@@ -53,6 +53,7 @@
     "react-table": "^7.8.0",
     "react-tweet-embed": "^2.0.0",
     "react-use": "^17.4.0",
+    "react-uuid": "^2.0.0",
     "react-window": "^1.8.9",
     "rehype-mathjax": "^4.0.2",
     "rehype-slug": "^5.1.0",
