+++ frontend/src/components/Notification/types.ts
@@ -1,0 +1,1 @@
+export interface Props {}
