+++ frontend/src/views/WheresMyGeneV2/components/GeneSearchBar/components/CitationButton/style.ts
@@ -1,8 +1,6 @@
 import styled from "@emotion/styled";
-import { ButtonIcon, fontBodyXs, fontBodyXxs, Icon } from "@czi-sds/components";
-import { NotificationWrapper } from "src/components/common/Filter/common/style";
-import { HEADER_HEIGHT_PX } from "src/components/Header/style";
-import { gray500, gray100, gray200 } from "src/common/theme";
+import { ButtonIcon, fontBodyXxs } from "@czi-sds/components";
+import { gray500 } from "src/common/theme";
 
 export const StyledButtonDiv = styled.div`
   display: flex;
@@ -21,34 +19,4 @@ export const StyledButtonIcon = styled(ButtonIcon)`
   width: 30px;
   height: 30px;
   top: 4px;
-`;
-
-export const StyledNotificationWrapper = styled(NotificationWrapper)`
-  top: ${HEADER_HEIGHT_PX}px;
-  right: 24px;
-`;
-
-export const StyledIcon = styled(Icon)`
-  height: 20px;
-  width: 20px;
-`;
-
-export const StyledNotificationLabel = styled.div`
-  ${fontBodyXs}
-
-  margin: 0 !important;
-  color: black;
-`;
-
-export const StyledNotificationDetails = styled.div`
-  ${fontBodyXs}
-
-  margin: 0 !important;
-  color: black;
-  width: 340px;
-  height: 120px;
-  padding: 8px 12px 8px 12px;
-  border: 1px solid ${gray200};
-  border-radius: 8px;
-  background-color: ${gray100};
 `;
