+++ frontend/src/views/WheresMyGeneV2/components/Main/index.tsx
@@ -34,8 +34,9 @@ import { CELL_INFO_SIDEBAR_WIDTH_PX } from "src/views/WheresMyGeneV2/components/
 import { UnderlyingDataChangeBanner } from "../GeneSearchBar/components/SaveExport/ExportBanner";
 import { GENE_EXPRESSION_BANNER_SURVEY_LINK } from "src/common/constants/airtableLinks";
 import { StyledRightSideBar } from "./style";
-import NormalizationNotification from "../GeneSearchBar/components/NormalizationNotification";
+// import NormalizationNotification from "../GeneSearchBar/components/NormalizationNotification";
 import { useConnect } from "src/views/WheresMyGeneV2/components/Main/connect";
+import Notification from "src/components/Notification";
 
 export const INFO_PANEL_WIDTH_PX = 320;
 
@@ -93,7 +94,7 @@ export default function WheresMyGene(): JSX.Element {
       <Head>
         <title>Gene Expression - CZ CELLxGENE Discover</title>
       </Head>
-      <NormalizationNotification />
+      {/* <NormalizationNotification /> */}
       <SideBar
         label="Filters"
         SideBarWrapperComponent={SideBarWrapper}
@@ -146,14 +147,12 @@ export default function WheresMyGene(): JSX.Element {
       <View id="view" overflow="hidden">
         <Wrapper>
           {isLoading && <Loader />}
-
           {/* Used for PNG and SVG exports to render message banner to render in output */}
           {downloadStatus.isLoading && (
             <StyledBannerContainer>
               <UnderlyingDataChangeBanner />
             </StyledBannerContainer>
           )}
-
           <Top id="top-legend">
             <Legend
               selectedCellTypes={cellTypesByTissueName}
@@ -169,8 +168,8 @@ export default function WheresMyGene(): JSX.Element {
               filteredCellTypes={filteredCellTypes}
               maxExpression={scaledMeanExpressionMax}
             />
+            <Notification />
           </Top>
-
           <StyledSidebarDrawer
             position="right"
             isOpen={isSourceDatasetSidebarOpen}
